move code to separate class
